backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
aiohttp==3.8.5
beautifulsoup4==4.12.2
pandas==2.0.3
sqlalchemy==2.0.21
google-api-python-client==2.100.0
tweepy==4.14.0
yfinance==0.2.21
selenium==4.13.0
python-dotenv==1.0.0
google-generativeai==0.4.1